refactor: rename Renderer and Page
