+++ src/components/Page/Page.stories.tsx
@@ -1,11 +1,18 @@
-import { Page } from "./Page.js";
-import PageObject from "./Page.json";
+import BlocksObject from "../Blocks/BlockList.json";
+import { OpenedHeading1 } from "../Blocks/Custom/OpenedHeading1.js";
+import { OpenedHeading2 } from "../Blocks/Custom/OpenedHeading2.js";
+import { OpenedHeading3 } from "../Blocks/Custom/OpenedHeading3.js";
+import { OpenedToggle } from "../Blocks/Custom/OpenedToggle.js";
 
-import type { PageObject as PageObjectType } from "../../types/notion/pages/page.js";
+import { Page } from "./Page.js";
+import PageMetaObjectData from "./PageMeta.json";
+
+import type { BlockBlockObject } from "../../types/notion/blocks/block.js";
+import type { PageObject } from "../../types/notion/pages/page.js";
 import type { Meta, StoryObj } from "@storybook/react";
 
 const meta: Meta<typeof Page> = {
-  title: "Page",
+  title: "Page/Page",
   component: Page,
   tags: ["autodocs"],
 };
@@ -16,26 +23,14 @@ type Story = StoryObj<typeof Page>;
 
 export const Default: Story = {
   args: {
-    page: PageObject as PageObjectType,
-  },
-  decorators: [
-    (Story) => (
-      <div className="notion-root notion-light">
-        <Story />
-      </div>
-    ),
-  ],
-};
-
-export const Dark: Story = {
-  args: {
-    page: PageObject as PageObjectType,
+    page: PageMetaObjectData as PageObject,
+    blocks: BlocksObject as Array<BlockBlockObject>,
     displayProperties: [
-      "USxG",
-      "Tags",
-      "Text",
-      "Multi-select",
       "Status",
+      "Multi-select",
+      "Text",
+      "Tags",
+      "Created time",
       "Last edited time",
       "Related Child database",
       "Select",
@@ -48,11 +43,25 @@ export const Dark: Story = {
       "Phone",
     ],
   },
-  decorators: [
-    (Story) => (
-      <div style={{ blockSize: "100vh" }} className="notion-root notion-dark">
-        <Story />
-      </div>
-    ),
-  ],
+};
+
+export const Dark: Story = {
+  args: {
+    page: PageMetaObjectData as PageObject,
+    blocks: BlocksObject as Array<BlockBlockObject>,
+    theme: "dark",
+  },
+};
+
+export const Custom: Story = {
+  args: {
+    page: PageMetaObjectData as PageObject,
+    blocks: BlocksObject as Array<BlockBlockObject>,
+    blockMapper: {
+      toggle: OpenedToggle,
+      heading_1: OpenedHeading1,
+      heading_2: OpenedHeading2,
+      heading_3: OpenedHeading3,
+    },
+  },
 };
